Bump version
This is needed because c3700b903107fd0b4b3ec95b2781dda32f2a11a5 changed
the code, but did *NOT* bump the version.  Causing people before the
merge having a 1.0 version of this module without the fix, and never
getting updated with zef because the version number didn't change.